requirements.txt
numpy
pandas
pillow
torch
transformers
tqdm